db diagram: translate to Sphinx extension
